Create a6-converterClass
Separate converter class from script file and export. Rename class and methods to allow for expansion to other data types.
